+++ src/app/(landing)/pricing/solutions/solutions_content.tsx
@@ -1,0 +1,13 @@
+"use client";
+
+import SolutionsPricingSection from "@/components/pricing/solutions_pricing_section";
+import { usePricing } from "@/contexts/pricing_context";
+
+export default function SolutionsContent() {
+  const { onSelectCustomPlan } = usePricing();
+  return (
+    <>
+      <SolutionsPricingSection onSelect={onSelectCustomPlan} />
+    </>
+  );
+}
